add rollup-plugin-visualizer, add manual chunks, some minor optimizations
