+++ website/configs/node.tsconfig.json
@@ -21,7 +21,7 @@
   },
 
   "include": [
-    "../vite.config.ts",
+    "../vite.config.mts",
     "../configs/**.ts"
   ]
 }
