+++ website/index.html
@@ -18,9 +18,9 @@
 
     <!-- Eclipse Cookie Consent -->
     <link rel="stylesheet" type="text/css"
-        href="//www.eclipse.org/eclipse.org-common/themes/solstice/public/stylesheets/vendor/cookieconsent/cookieconsent.min.css" />
+        href="https://www.eclipse.org/eclipse.org-common/themes/solstice/public/stylesheets/vendor/cookieconsent/cookieconsent.min.css" />
     <script
-        src="//www.eclipse.org/eclipse.org-common/themes/solstice/public/javascript/vendor/cookieconsent/default.min.js">
+        src="https://www.eclipse.org/eclipse.org-common/themes/solstice/public/javascript/vendor/cookieconsent/default.min.js" type="">
     </script>
     <style>
         a.cc-btn.cc-allow {
