+++ website/package.json
@@ -29,6 +29,7 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-react": "^7.37.0",
     "prettier": "^3.8.1",
+    "rollup-plugin-visualizer": "^7.0.1",
     "typescript": "^5.9.0",
     "typescript-eslint": "^8.57.0",
     "vite": "^7.3.0",
@@ -36,8 +37,8 @@
   },
   "scripts": {
     "prepare": "yarn build",
-    "format": "prettier --write --config .prettierrc src/**/*.{ts,tsx,mjs}",
-    "lint": "prettier --check --config .prettierrc src/**/*.{ts,tsx,mjs} && eslint -c eslint.config.mjs src",
+    "format": "prettier --write --config .prettierrc src/**/*.{ts,tsx,mjs,mts}",
+    "lint": "prettier --check --config .prettierrc src/**/*.{ts,tsx,mjs,mts} && eslint -c eslint.config.mjs src",
     "build": "tsc -p ./configs/app.tsconfig.json && vite build",
     "dev": "vite dev",
     "preview": "vite preview"
