+++ website/vite.config.mts
@@ -1,0 +1,45 @@
+import path from 'node:path';
+import react from '@vitejs/plugin-react';
+import webfontDownload from 'vite-plugin-webfont-dl';
+import { visualizer } from 'rollup-plugin-visualizer';
+import { defineConfig, type PluginOption } from 'vite';
+
+const outRootDir = path.join(__dirname, 'dist');
+
+export default defineConfig(() => ({
+  plugins: [react(), webfontDownload(), visualizer() as PluginOption],
+  server: {
+    host: true,
+    port: 3000
+  },
+  preview: {
+    port: 3000
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src')
+    }
+  },
+  publicDir: 'static',
+  build: {
+    target: 'es2020',
+    minify: true,
+    sourcemap: true,
+    outDir: outRootDir,
+    emptyOutDir: true,
+    chunkSizeWarningLimit: 800,
+    rollupOptions: {
+      output: {
+        entryFileNames: 'bundle-[hash].js',
+        assetFileNames: 'bundle-[name].css',
+        manualChunks: {
+          lodash: ['lodash'],
+          luxon: ['luxon'],
+          react: ['react', 'react-router-dom', 'react-dom'],
+          material: ['@mui/material'],
+          'mui-xtnd': ['@mui/x-charts', '@mui/x-data-grid', '@mui/x-date-pickers'],
+        },
+      }
+    }
+  }
+}));
