--- website/vite.config.ts
@@ -1,36 +1,0 @@
-import path from 'node:path';
-import react from '@vitejs/plugin-react';
-import webfontDownload from 'vite-plugin-webfont-dl';
-import { defineConfig } from 'vite';
-
-const outRootDir = path.join(__dirname, 'dist');
-
-export default defineConfig(() => ({
-  plugins: [react(), webfontDownload()],
-  server: {
-    host: true,
-    port: 3000
-  },
-  preview: {
-    port: 3000
-  },
-  resolve: {
-    alias: {
-      '@': path.resolve(__dirname, './src')
-    }
-  },
-  publicDir: 'static',
-  build: {
-    target: 'es2020',
-    minify: true,
-    sourcemap: true,
-    outDir: outRootDir,
-    emptyOutDir: true,
-    rollupOptions: {
-      output: {
-        entryFileNames: 'bundle.js',
-        assetFileNames: 'bundle-[name].css'
-      }
-    }
-  }
-}));
